Added Interface AbstractActionIU and new class Point

+++ src/pentogame/inproObjects/ActionStrength.java
@@ -3,14 +3,14 @@ package pentogame.inproObjects;
 public enum ActionStrength {
 	NONE, WEAK, NORMAL, STRONG, MAX;
 	
-	double getDistance() {
+	int getDistance() {
 		switch (this) {
-		case NONE: return 0.0;
-		case WEAK: return 1.0;
-		case NORMAL: return 3.0;
-		case STRONG: return 10.0;
-		case MAX: return 99999.0; // this prints more favourably than Double.MAX_VALUE
-		default: return 0.0;
+		case NONE: return 0;
+		case WEAK: return 1;
+		case NORMAL: return 3;
+		case STRONG: return 10;
+		case MAX: return 99999; // this prints more favourably than Double.MAX_VALUE
+		default: return 0;
 		}
 	}
